Tests/LibWeb/Text/expected/css/custom-property-computed-value.txt

"Hello" "world!"
supports-valid
supports-fallback
media-valid
media-fallback
"fallback "
